gcloud/functions/requirements.txt
functions-framework==3.*
google-cloud-firestore
numpy